pyproject.toml
[tool.poetry]
name = "page-stream-segmentation"
version = "0.1.0"
description = ""
authors = ["Alexander Giagoulas <alexander.giagoulas@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-multipart = "^0.0.5"
pdf2image = "^1.14.0"
uvicorn = "^0.13.4"
fastapi = "^0.63.0"
pytesseract = "^0.3.7"
Pillow = "^8.1.2"
fasttext = "^0.9.2"
scipy = "^1.6.1"
scikit-learn = "^0.24.1"
Keras = "^2.4.3"
tensorflow = "^2.4.1"
argparse = "^1.4.0"
PyPDF2 = "^1.26.0"
opencv-python = "^4.5.1"
fpdf = "^1.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
